reactor to single page
