+++ src/frontend/komponenter/Oppgaver/VisOppgaver.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import FilterSkjema from './FilterSkjema';
+import OppgaveList from './OppgaveList';
+
+const VisOppgaver: React.FunctionComponent = () => {
+    return (
+        <div className="visoppgaver">
+            <FilterSkjema />
+            <OppgaveList />
+        </div>
+    );
+};
+
+export default VisOppgaver;
